Fine tuning

+++ HaddySimHub/Models/RaceData.cs
@@ -141,7 +141,7 @@ public record TimingEntry
     public string LicenseColor { get; init; } = string.Empty;
     public long IRating { get; init; }
     public int Laps { get; init; }
-    public float LapCompletedPct { get; init; }
+    public int LapCompletedPct { get; init; }
     public bool IsPlayer { get; init; }
     public bool IsSafetyCar { get; init; }
     public bool IsInPits { get; init; }
